+++ src/modules/CMakeLists.txt
@@ -31,5 +31,8 @@ IF(MSVC)
 ENDIF()
 
 #ADD_SUBDIRECTORY(lua)
-#ADD_SUBDIRECTORY(python)
+IF(Python3_FOUND)
+    ADD_SUBDIRECTORY(python)
+ENDIF()
+
 ADD_SUBDIRECTORY(stb_image)
