+++ src/modules/python.old/test/test002_inherit.py
